Add support for Dallas 18B20 Temp Sensor via MQTT
This commit is somewhat dirty. I used the original IRMQTTServer
which came with Arduino IDE and didn't start from the latest github version.

As such this commit is a bit gross, and I can't be bothered installing a newer version
on the ESP8266 when this code currently works and does what I want.